Imported inline documentation for use with the builtin help function.
To make this work better, I have removed tp_getattr in favor of tp_getattro 
and tp_methods.

This method has much better support for inline help, but relies on 
PyObject_GenericGetAttr.

From trawling through the python branches, it looks as if 
PyObject_GenericGetAttr was introduced in Python 2.2.

I think it is an acceptable to trade better inline documentation with no
support for versions earlier than 2.2.

(I *could* do both, but my time is finite).